Removed BitmapConverter class as this relied on Windows-only code. This eliminates the dependency on System.Drawing and thus GDI+. The converter class has been moved into the FitsViewer sample application for now. A better long-term strategy might involve use of a cross-platform graphics library such as ImageSharp.
