+++ TA.ObjectOrientedAstronomy.FlexibleImageTransportSystem/TA.ObjectOrientedAstronomy.FlexibleImageTransportSystem.csproj
@@ -27,7 +27,6 @@
     <ItemGroup>
       <PackageReference Include="JetBrains.Annotations" Version="2020.1.0" />
       <PackageReference Include="NLog" Version="4.7.3" />
-      <PackageReference Include="System.Drawing.Common" Version="4.7.0" />
       <PackageReference Include="TA.Utils.Core" Version="1.1.1" />
     </ItemGroup>
     <ItemGroup>
