+++ TA.ObjectOrientedAstronomy.Specifications/FitsReaderSpecs.cs
@@ -13,6 +13,7 @@ using Machine.Specifications;
 using TA.ObjectOrientedAstronomy.FlexibleImageTransportSystem;
 using TA.ObjectOrientedAstronomy.FlexibleImageTransportSystem.PropertyBinder;
 using TA.ObjectOrientedAstronomy.Specifications.Builders;
+using TA.ObjectOrientedAstronomy.Specifications.TestHelpers;
 
 // ReSharper disable ComplexConditionExpression
 
@@ -169,24 +170,6 @@ namespace TA.ObjectOrientedAstronomy.Specifications
             () => hdu.DataArrayLengthBytes.ShouldEqual(hdu.DataArrayLengthBits / 8);
         It should_populate_the_data_array = () => hdu.RawData.Length.ShouldEqual(hdu.DataArrayLengthBytes);
         It should_set_the_data_type_to_image = () => hdu.DataType.ShouldEqual(FitsDataType.Image);
-        static FitsHeaderDataUnit hdu;
-        }
-
-    [Subject(typeof(FitsReader))]
-    class when_converting_monocrome_image_data_to_a_windows_bitmap : with_fits_reader
-        {
-        Establish context = () =>
-            {
-            FitsReader = ContextBuilder.FromEmbeddedResource("WFPC2ASSNu5780205bx.fits").Build();
-            hdu = FitsReader.ReadPrimaryHeaderDataUnit().WaitFoResult();
-            };
-        Because of = () => bitmap = hdu.ToWindowsBitmap();
-        It should_produce_a_bitmap_with_the_expected_x_dimension =
-            () => bitmap.Width.ShouldEqual(hdu.MandatoryKeywords.LengthOfAxis[0]);
-        It should_produce_a_bitmap_with_the_expected_y_dimension =
-            () => bitmap.Height.ShouldEqual(hdu.MandatoryKeywords.LengthOfAxis[1]);
-        It should_produce_an_rgb_24bpp_image = () => bitmap.PixelFormat.ShouldEqual(PixelFormat.Format24bppRgb);
-        static Bitmap bitmap;
         static FitsHeaderDataUnit hdu;
         }
     }
